+++ CodeStarterCamp_Week3/Coffee.swift
@@ -1,0 +1,18 @@
+//
+//  Coffee.swift
+//  CodeStarterCamp_Week3
+//
+//  Created by 김형철 on 2023/03/09.
+//
+
+import Foundation
+
+
+
+enum Coffee {
+    case latte
+    case americano
+    case lemonade
+}
+
+
