+++ CodeStarterCamp_Week3/main.swift
@@ -8,47 +8,7 @@
 
 import Foundation
 
-
-
-struct Person {
-    var money: Int
-    
-    mutating func buyCoffee(coffeePrice: CoffeShop, coffee: Coffee) {
-        
-        if let price = coffeePrice.menu[coffee] {
-            self.money -= price
-        }
-
-    }
-}
-
-
-struct CoffeShop {
-    var barista: Person!
-    var sales: Int
-    var menu: Dictionary<Coffee, Int>
-    var pickUpTable: [Coffee]
-    
-    mutating func takeOrder(for coffee: Coffee) -> Coffee {
-        if coffee == Coffee.latte {
-            self.pickUpTable.append(coffee)
-        } else if coffee == .americano {
-            self.pickUpTable.append(coffee)
-        } else {
-            self.pickUpTable.append(coffee)
-        }
-        
-        return pickUpTable[pickUpTable.endIndex - 1]
-    }
-    
-    
-}
-
-enum Coffee {
-    case latte
-    case americano
-    case lemonade
-}
+typealias CoffeePrice = Int
 
 var misterLee: Person = Person(money: 1000)
 var missKim: Person = Person(money: 2000)
